+++ python/dask_cudf/dask_cudf/io/tests/test_parquet.py
@@ -153,3 +153,27 @@ def test_filters(tmpdir):
     )
     assert c.npartitions <= 1
     assert not len(c)
+
+
+@pytest.mark.parametrize(
+    "parts", [["year", "month", "day"], ["year", "month"], ["year"]]
+)
+def test_roundtrip_from_dask_partitioned(tmpdir, parts):
+    tmpdir = str(tmpdir)
+
+    df = pd.DataFrame()
+    df["year"] = [2018, 2019, 2019, 2019, 2020, 2021]
+    df["month"] = [1, 2, 3, 3, 3, 2]
+    df["day"] = [1, 1, 1, 2, 2, 1]
+    df["data"] = [0, 0, 0, 0, 0, 0]
+    ddf2 = dd.from_pandas(df, npartitions=2)
+
+    ddf2.to_parquet(tmpdir, engine="pyarrow", partition_on=parts)
+    df_read = dd.read_parquet(tmpdir).compute()
+    gdf_read = dask_cudf.read_parquet(tmpdir).compute()
+    # dask_cudf converts categoricals to original data type,
+    # dask.dataframe does not. Convert types before check...
+    for part in parts:
+        cat_type = df_read[part].dtype._categories.dtype
+        df_read[part] = df_read[part].astype(cat_type)
+    assert_eq(df_read, gdf_read)